feat: implement authentication functionality
