+++ package.json
@@ -11,11 +11,8 @@
     "preview": "vite preview"
   },
   "dependencies": {
-    "@emotion/react": "^11.11.1",
-    "@emotion/styled": "^11.11.0",
     "@mui/icons-material": "^5.11.16",
     "@mui/lab": "^5.0.0-alpha.133",
-    "@mui/material": "^5.13.4",
     "@testing-library/react": "^14.0.0",
     "@types/moment": "^2.13.0",
     "connect": "^3.7.0",
@@ -24,6 +21,9 @@
     "react-dom": "^18.2.0"
   },
   "devDependencies": {
+    "@emotion/react": "^11.11.1",
+    "@emotion/styled": "^11.11.0",
+    "@mui/material": "^5.14.5",
     "@testing-library/jest-dom": "^5.16.5",
     "@types/node": "^20.2.5",
     "@types/react": "^18.0.37",
